+++ src/components/Ranking.jsx
@@ -1,0 +1,70 @@
+import React, { useState, useEffect } from 'react';
+import './Ranking.scss';
+
+const Ranking = ({ rankType = null, itemsToShow = 5, valueLabels = null }) => {
+  const [rankData, setRankData] = useState(null); 
+  
+  if(!rankType) return(null);
+
+  useEffect(() => {
+    const fetchRanking = async()  => {  
+      try {
+        const response = await fetch(`${import.meta.env.VITE_CLOUD_URL}/mainframe/ranking`,
+          {
+            method: 'POST',
+            headers: {
+              "x-api-key": import.meta.env.VITE_CLOUD_APIKEY,
+              "Content-Type": "application/json"
+            },
+            body: JSON.stringify({
+              rank_type: rankType, items_to_show: itemsToShow
+            })
+          });
+        if(response) {
+          const result = await response.json();
+          setRankData(result);
+        }
+      } catch(e) {
+        console.log('[Ranking] Error: ' + e.message);
+      }
+    };
+
+    fetchRanking();
+
+  },[]);
+
+  return (
+    <div className="grid-ranking">
+      {!rankData && (
+        <p>Ranking data is currently unavailable.</p>
+      )}
+      {rankData && rankData.map((rankItem,idx) => (
+        <div className="row" key={idx}>
+          <div className="rank">#{idx+1}</div>
+          <div className="icon">
+            {rankItem.twitch_avatar ? (
+              <div className={'avatar user-level-bg level-'+rankItem.level}>
+                <img src={rankItem.twitch_avatar} alt="Avatar" />
+              </div>
+            ) : (
+              <i className={'fa-solid fa-user user-level level-'+rankItem.level}></i>
+            )}
+            
+          </div>
+          <div className="player">
+            <h4>{rankItem.twitch_display_name}</h4>
+            <div className="data"><span className="level">Lvl {rankItem.level}</span> <span className={'title user-level level-'+rankItem.level}>{rankItem.title}</span></div>
+          </div>
+          {valueLabels && (
+            <div className="points">
+              <span className={'user-level level-'+rankItem.level}>{Math.ceil(rankItem.value)}</span>
+              <span className="value-label">{valueLabels}</span>
+            </div>
+          )}
+        </div>
+      ))}
+    </div>
+  )
+}
+
+export default Ranking